fix(subscribers): focus the affordance a failed read renders
